hf_backend/requirements.txt
fastapi
uvicorn
transformers
torch
python-jose
passlib
bcrypt==3.2.2
sqlite-utils
